rename: isChelseaBridgeNotification.ts -> notifications.ts
